+++ netpyne/analysis.py
@@ -218,7 +218,8 @@ def getCellsInclude(include):
 ## Raster plot 
 ######################################################################################################################################################
 def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, orderBy = 'gid', orderInverse = False, labels = 'legend', popRates = False,
-        spikeHist = None, spikeHistBin = 5, syncLines = False, lw = 2, marker = '|', figSize = (10,8), saveData = None, saveFig = None, showFig = True): 
+        spikeHist = None, spikeHistBin = 5, syncLines = False, lw = 2, marker = '|', popColors = None, figSize = (10,8), saveData = None, saveFig = None, 
+        showFig = True): 
     ''' 
     Raster plot of network cells 
         - include (['all',|'allCells',|'allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): Cells to include (default: 'allCells')
@@ -233,6 +234,7 @@ def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, order
         - syncLines (True|False): calculate synchorny measure and plot vertical lines for each spike to evidence synchrony (default: False)
         - lw (integer): Line width for each spike (default: 2)
         - marker (char): Marker for each spike (default: '|')
+        - popColors (dict): Dictionary with color (value) used for each population (key) (default: None)
         - figSize ((width, height)): Size of figure (default: (10,8))
         - saveData (None|True|'fileName'): File name where to save the final data used to generate the figure; 
             if set to True uses filename from simConfig (default: None)
@@ -256,7 +258,9 @@ def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, order
     selectedPops = [cell['tags']['popLabel'] for cell in cells]
     popLabels = [pop for pop in sim.net.allPops if pop in selectedPops] # preserves original ordering
     if netStimLabels: popLabels.append('NetStims')
-    popColors = {popLabel: colorList[ipop%len(colorList)] for ipop,popLabel in enumerate(popLabels)} # dict with color for each pop
+    popColorsTmp = {popLabel: colorList[ipop%len(colorList)] for ipop,popLabel in enumerate(popLabels)} # dict with color for each pop
+    if popColors: popColorsTmp.update(popColors)
+    popColors = popColorsTmp
     if len(cellGids) > 0:
         gidColors = {cell['gid']: popColors[cell['tags']['popLabel']] for cell in cells}  # dict with color for each gid
         try:
@@ -462,7 +466,7 @@ def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, order
 ## Plot spike histogram
 ######################################################################################################################################################
 def plotSpikeHist (include = ['allCells', 'eachPop'], timeRange = None, binSize = 5, overlay=True, graphType='line', yaxis = 'rate', 
-    figSize = (10,8), saveData = None, saveFig = None, showFig = True): 
+    popColors = None, figSize = (10,8), saveData = None, saveFig = None, showFig = True): 
     ''' 
     Plot spike histogram
         - include (['all',|'allCells','allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): List of data series to include. 
@@ -472,6 +476,7 @@ def plotSpikeHist (include = ['allCells', 'eachPop'], timeRange = None, binSize 
         - overlay (True|False): Whether to overlay the data lines or plot in separate subplots (default: True)
         - graphType ('line'|'bar'): Type of graph to use (line graph or bar plot) (default: 'line')
         - yaxis ('rate'|'count'): Units of y axis (firing rate in Hz, or spike count) (default: 'rate')
+        - popColors (dict): TO DO!
         - figSize ((width, height)): Size of figure (default: (10,8))
         - saveData (None|True|'fileName'): File name where to save the final data used to generate the figure;
             if set to True uses filename from simConfig (default: None)
@@ -752,7 +757,7 @@ def plotRatePSD (include = ['allCells', 'eachPop'], timeRange = None, binSize = 
 ######################################################################################################################################################
 ## Plot recorded cell traces (V, i, g, etc.)
 ######################################################################################################################################################
-def plotTraces (include = None, timeRange = None, overlay = False, oneFigPer = 'cell', rerun = False,
+def plotTraces (include = None, timeRange = None, overlay = False, oneFigPer = 'cell', rerun = False, colors = None,
     figSize = (10,8), saveData = None, saveFig = None, showFig = True): 
     ''' 
     Plot recorded traces
@@ -763,6 +768,7 @@ def plotTraces (include = None, timeRange = None, overlay = False, oneFigPer = '
         - oneFigPer ('cell'|'trace'): Whether to plot one figure per cell (showing multiple traces) 
             or per trace (showing multiple cells) (default: 'cell')
         - rerun (True|False): rerun simulation so new set of cells gets recorded (default: False)
+        - colors (list): List of normalized RGB colors to use for traces
         - figSize ((width, height)): Size of figure (default: (10,8))
         - saveData (None|True|'fileName'): File name where to save the final data used to generate the figure; 
             if set to True uses filename from simConfig (default: None)
@@ -776,6 +782,10 @@ def plotTraces (include = None, timeRange = None, overlay = False, oneFigPer = '
     print('Plotting recorded cell traces ...')
 
     if include is None: include = [] # If not defined, initialize as empty list
+    if isinstance(colors, list): 
+        colorList = colors
+    else: 
+        global colorList
 
     # rerun simulation so new include cells get recorded from
     if rerun: 
@@ -818,11 +828,12 @@ def plotTraces (include = None, timeRange = None, overlay = False, oneFigPer = '
                     plt.plot(t[:len(data)], data, linewidth=1.5, color=color, label='Cell %d, Pop %s '%(int(gid), gidPops[gid]))
                     plt.xlabel('Time (ms)', fontsize=fontsiz)
                     plt.xlim(timeRange)
-                    plt.title('Cell %d, Pop %s '%(int(gid), gidPops[gid]))
+                    plt.title(trace)
             if overlay:
-                maxLabelLen = 10
-                plt.subplots_adjust(right=(0.9-0.012*maxLabelLen)) 
-                plt.legend(fontsize=fontsiz, bbox_to_anchor=(1.04, 1), loc=2, borderaxespad=0.)
+                #maxLabelLen = 10
+                #plt.subplots_adjust(right=(0.9-0.012*maxLabelLen)) 
+                #plt.legend(fontsize=fontsiz, bbox_to_anchor=(1.04, 1), loc=2, borderaxespad=0.)
+                plt.legend()
 
     # Plot one fig per cell
     if oneFigPer == 'cell':
@@ -1493,13 +1504,14 @@ def plotConn (includePre = ['all'], includePost = ['all'], feature = 'strength',
 ######################################################################################################################################################
 ## Plot 2D representation of network cell positions and connections
 ######################################################################################################################################################
-def plot2Dnet (include = ['allCells'], figSize = (12,12), view = 'xy', showConns = True, saveData = None, saveFig = None, showFig = True): 
+def plot2Dnet (include = ['allCells'], figSize = (12,12), view = 'xy', showConns = True, popColors = None, saveData = None, saveFig = None, showFig = True): 
     ''' 
     Plot 2D representation of network cell positions and connections
         - include (['all',|'allCells','allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): Cells to show (default: ['all'])
         - showConns (True|False): Whether to show connections or not (default: True)
         - figSize ((width, height)): Size of figure (default: (12,12))
         - view ('xy', 'xz'): Perspective view: front ('xy') or top-down ('xz')
+        - popColors (dict): Dictionary with color (value) used for each population (key) (default: None)
         - saveData (None|'fileName'): File name where to save the final data used to generate the figure (default: None)
         - saveFig (None|'fileName'): File name where to save the figure;
             if set to True uses filename from simConfig (default: None)(default: None)
@@ -1521,7 +1533,9 @@ def plot2Dnet (include = ['allCells'], figSize = (12,12), view = 'xy', showConns
     cells, cellGids, _ = getCellsInclude(include)           
     selectedPops = [cell['tags']['popLabel'] for cell in cells]
     popLabels = [pop for pop in sim.net.allPops if pop in selectedPops] # preserves original ordering
-    popColors = {popLabel: colorList[ipop%len(colorList)] for ipop,popLabel in enumerate(popLabels)} # dict with color for each pop
+    popColorsTmp = {popLabel: colorList[ipop%len(colorList)] for ipop,popLabel in enumerate(popLabels)} # dict with color for each pop
+    if popColors: popColorsTmp.update(popColors)
+    popColors = popColorsTmp
     cellColors = [popColors[cell['tags']['popLabel']] for cell in cells]
 
     # front view
@@ -1818,6 +1832,74 @@ def granger(cells1 = [], cells2 = [], spks1 = None, spks2 = None, label1 = 'spkT
         if showFig: _showFigure()
 
     return F, Fx2y[0],Fy2x[0], Fxy[0], fig
+
+
+
+######################################################################################################################################################
+## EPSPs amplitude
+######################################################################################################################################################
+def plotEPSPAmp(include=None, trace=None, start=0, interval=50, number=2, amp='absolute', saveFig=False, showFig=True):
+
+    print('Plotting EPSP amplitudes...')
+
+    if include is None: include = [] # If not defined, initialize as empty list
+
+    cells, cellGids, _ = getCellsInclude(include)
+    gidPops = {cell['gid']: cell['tags']['popLabel'] for cell in cells}
+
+    if not trace: 
+        print 'Error: Missing trace to to plot EPSP amplitudes'
+        return
+    step = sim.cfg.recordStep
+
+    peaksAbs = np.zeros((number, len(cellGids)))
+    peaksRel = np.zeros((number, len(cellGids)))
+    for icell, gid in enumerate(cellGids):
+        vsoma = sim.allSimData[trace]['cell_'+str(gid)]
+        for ipeak in range(number):
+            peakAbs = max(vsoma[int(start/step+(ipeak*interval/step)):int(start/step+(ipeak*interval/step)+(interval-1)/step)]) 
+            peakRel = peakAbs - vsoma[int((start-1)/step)]
+            peaksAbs[ipeak,icell] = peakAbs
+            peaksRel[ipeak,icell] = peakRel
+
+    if amp == 'absolute':
+        peaks = peaksAbs
+        ylabel = 'EPSP peak V (mV)'
+    elif amp == 'relative':
+        peaks = peaksRel
+        ylabel = 'EPSP amplitude (mV)'
+    elif amp == 'ratio':
+        peaks = np.zeros((number, len(cellGids)))
+        for icell in range(len(cellGids)):
+            peaks[:, icell] = peaksRel[:, icell] / peaksRel[0, icell]
+        ylabel = 'EPSP amplitude ratio'
+        
+    xlabel = 'EPSP number'
+
+    # plot
+    fig = plt.figure()
+    plt.plot(peaks, marker='o')
+    plt.xlabel(xlabel)
+    plt.ylabel(ylabel)
+    h = plt.axes()
+    ylim=list(h.get_ylim())
+    h.set_ylim(ylim[0], ylim[1]+(0.15*abs(ylim[1]-ylim[0])))
+    h.set_xticks(range(number))
+    h.set_xticklabels(range(1, number+1))
+    plt.legend(gidPops.values())
+
+    # save figure
+    if saveFig: 
+        if isinstance(saveFig, basestring):
+            filename = saveFig
+        else:
+            filename = sim.cfg.filename+'_'+'EPSPamp_'+amp+'.png'
+        plt.savefig(filename)
+
+    # show fig 
+    if showFig: _showFigure()
+
+    return peaks, fig